Update About.xml
ExportedXenotypeStorageModFolder seems to stay high up, even with  rimpy "put as low as possible". on. So, I have it so the pre-set will always put it bellow all mods. Also put the xmlpatchhelper, as, while not necessary, it's nice to know it is guarenteed to stay bellow So, I told it to always put this mod first.